refactor commands and finish dropping core
