+++ src/kimera_eval/__init__.py
@@ -1,6 +1,9 @@
 """Main entry point for running evaluations."""
 import kimera_eval.logger
-import kimera_eval.tools.cli
+import kimera_eval.tools.evaluate
+import kimera_eval.tools.run
+import kimera_eval.tools.summary
+import kimera_eval.tools.timing
 
 from kimera_eval.trajectory_metrics import *
 from kimera_eval.experiment_config import *
@@ -16,4 +19,7 @@ def main(log_level):
     kimera_eval.logger.configure_logging(level=log_level)
 
 
-main.add_command(kimera_eval.tools.cli.run)
+main.add_command(kimera_eval.tools.run.run)
+main.add_command(kimera_eval.tools.evaluate.run)
+main.add_command(kimera_eval.tools.timing.run)
+main.add_command(kimera_eval.tools.summary.run)
